# Optimize GetMD5 function