--- a/MediaBrowser.Common/Extensions/BaseExtensions.cs
+++ b/MediaBrowser.Common/Extensions/BaseExtensions.cs
@@ -31,11 +31,7 @@ namespace MediaBrowser.Common.Extensions
         public static Guid GetMD5(this string str)
         {
 #pragma warning disable CA5351
-            using (var provider = MD5.Create())
-            {
-                return new Guid(provider.ComputeHash(Encoding.Unicode.GetBytes(str)));
-            }
-
+            return new Guid(MD5.HashData(Encoding.Unicode.GetBytes(str)));
 #pragma warning restore CA5351
         }
     }